% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_altitude.R
\name{get_altitude}
\alias{get_altitude}
\title{Get altitude from coordinates}
\usage{
get_altitude(.Data, .longitude = NULL, .latitude = NULL,
  .src = c("racemap", "geonames", "google", "openelevation"),
  .geonames.username = "demo", .google.api.key = NULL)
}
\arguments{
\item{.Data}{data.table or data.frame with columns \code{latitude} and \code{longitude}.
Either specify the \code{.Data} argument or specify \code{.longitude} and \code{.latitude}.}

\item{.longitude}{Numeric vector.}

\item{.latitude}{Numeric vector.}

\item{.src}{Character value. Which API service to use?
One of \code{c("geonames", "mapzen", "google", "openelevation")}.}

\item{.geonames.username}{Character value.}

\item{.google.api.key}{Character value.}
}
\value{
Numeric vector of altitudes.
}
\description{
This function returns the altitude (elevation) in meters for coordinates.
}
\details{
Currently different services can be used:
\itemize{
\item \code{geonames}: Uses the geonames webservice. The number of free requests is limited,
for a higher rate, register and pass your username to this function.
Check usage terms at \url{http://www.geonames.org/export/web-services.html}.
\item \code{google}: Uses the Google Elevation API.
You need to pass your API key. Costs can apply.
Check usage terms at
\url{https://developers.google.com/maps/documentation/elevation/usage-and-billing}.
\item \code{openelevation}: Uses free Open-Elevation API.
See \url{https://github.com/Jorl17/open-elevation}.
\item \code{racemap}: Uses free Racemap API. Based on mapzen terrain data.
See \url{https://github.com/racemap/elevation-service}.
}

It is recommended to use environment variables to pass secrets like your geonames
username or API keys to the function.

Note that some APIs do not return values for ocean areas, e.g. at
geonames those values are encoded as -32768.
}
\examples{
\dontrun{
Data <- data.frame(latitude = 48:50, longitude = 11:13)

# Geonames
get_altitude(Data, .src = "geonames")

# Set your geonames username as environment variable in .Renviron file before.
# get_altitude(Data, .src = "geonames", .geonames.username = Sys.getenv("GEONAMES_USERNAME"))

# Google
# get_altitude(Data, Data, .src = "google", .google.api.key = Sys.getenv("GOOGLE_API_KEY"))

# Racemap
get_altitude(Data, .src = "racemap")

# Open-Elevation
get_altitude(Data, .src = "openelevation")
}
}
